+++ BambooTracker/instrument/effect_iterator.hpp
@@ -12,6 +12,7 @@ public:
 	int getCommandData() const override;
 	int next(bool isReleaseBegin = false) override;
 	int front() override;
+	int end() override;
 
 private:
 	int pos_;
@@ -28,6 +29,7 @@ public:
 	int getCommandData() const override;
 	int next(bool isReleaseBegin = false) override;
 	int front() override;
+	int end() override;
 
 private:
 	int pos_;
@@ -44,6 +46,7 @@ public:
 	int getCommandData() const override;
 	int next(bool isReleaseBegin = false) override;
 	int front() override;
+	int end() override;
 
 private:
 	int pos_;